Move compiler arg parsing stuff to scrapscript.py
This makes it easier to bundle the compiler with a Cosmopolitan file and
also centralizes all the arg parsing stuff.
